requirements.txt
Flask
pywhatkit
gunicorn
pyttsx3
pywhatkit
datetime
wikipedia
pyjokes
SpeechRecognition
requests
python-dotenv
gTTs